+++ widgets/views/push-notification-info.php
@@ -1,3 +1,0 @@
-<div class="alert alert-danger">
-    <strong>Push Notifications</strong> not allowed.
-</div>